quoter_frontend/pyproject.toml
[project]
name = "quoter_frontend"
version = "0.1"
dependencies = [
   "requests==2.31.0",
   "streamlit==1.29.0",
   "streamlit-pdf-viewer",
   "python-dotenv",
]

[project.optional-dependencies]
dev = ["black", "flake8"]
